refactor: egctl cmds (#1167)
refactor egctl cmds.
